Overhauled Broker.java to now store two arrays, one of coin names (symbols) and the other of coin prices, in which both arrays are linked through index. Removed dependence in Broker.java on Coin objects. Switched over all getters and setters.
Added updating coin price functionality to MainUI.java inside ActionListener, as well as initializing each Broker entered by the user and executing each trade for the TradingStrategy linked to each Broker. Also added creating and initialization of TradingStrategies to MainUI.java.

Added BrokerName parameter to TradingStrategy.java method "Trade" so it can now return a trade result string which has all information to be graphed and added to the table.

Signed-off-by: SamuelCatania <69449284+SamuelCatania@users.noreply.github.com>